Update Titanium_SandBox Teamplates
Updated the template layout so that esthetically it looks nice when you are writing and testing code from within the SandBox. Live remote editing is a wonderful tool to use with this. I use Dreamweaver CS6 to edit my files and I have it set up so that when I save a file locally it automatically uploads it to the server. You can do this as well with MS Visual Code I believe.