Removed Sommaire
Sommaire has been removed as it is podunk, and should have been upgraded fixed or modified by now! To me it is garbage... Who would let a vital part of a CMS set for years as if it was just ok. It's not and that version is garbage! So we decided to 86it!